make differentiation test folder + fix adjoint accordingly + chex on adjoint test
